stemming frequency addition fix
when adding +1 to stemming word frequency, only add 1 to the stemmed word frequency only if the stemmed word is not equal to the original word (otherwise previous code was adding +2 to the word). 

EX: word = have, stemmed word = have. Only add +1 to have. word = having, stemmed word = have, add + 1 to having & + 1 to have.